searchとstatus以外は修正

+++ app/services/delete_group.rb
@@ -5,21 +5,17 @@ module Yuzakan
     class DeleteGroup < Yuzakan::ServiceOperation
       category :group
 
-      def call(groupname, services = nil)
-        groupname = step validate_name(groupname)
-        services = step get_services(services, method: :group_delete)
+      def call(service, groupname)
+        return unless can_call?(service, :group_delete)
 
-        services.to_h do |service|
-          adapter = step get_adapter(service)
-          groupdata = adapter.delete_group(groupname)
-          result =
-            if groupdata
-              cache_delete(service) # delete list
-              cache_delete(service, groupname)
-              step convert_data(service, groupdata)
-            end
-          [service.name, result]
-        end.compact
+        groupname = step validate_name(groupname)
+        adapter = step get_adapter(service)
+        groupdata = adapter.delete_group(groupname)
+        return unless groupdata
+
+        cache_delete(service) # delete list
+        cache_delete(service, groupname)
+        step convert_data(service, groupdata)
       end
     end
   end
